# Conflicts:
#	src/app/menubar/components/menubar-container/menubar-container.component.html

+++ src/app/classroom/overzicht/classroom/classroom.component.ts
@@ -1,8 +1,8 @@
 import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
-import {ProductTemplate} from "../../../types/productTemplate.model";
-import {TemplateService} from "../../../services/template.service";
-import {Classroom} from "../../../types/classroom.model";
-import {ClassroomService} from "../../../services/classroom.service";
+import {ProductTemplate} from '../../../types/productTemplate.model';
+import {TemplateService} from '../../../services/template.service';
+import {Classroom} from '../../../types/classroom.model';
+import {ClassroomService} from '../../../services/classroom.service';
 
 @Component({
   selector: 'app-classroom',
@@ -12,18 +12,17 @@ import {ClassroomService} from "../../../services/classroom.service";
 export class ClassroomComponent implements OnInit {
 
   @Input() public classroom: Classroom;
-  @Output() public deletedClassroom = new EventEmitter<Classroom>();
 
   constructor(private classroomService: ClassroomService) { }
 
   ngOnInit() {
   }
 
-  deleteClassroom(): boolean {
-    this.deletedClassroom.emit(this.classroom);
-    console.log(this.deletedClassroom);
-    this.classroomService.deleteClassroom(this.classroom).subscribe();
-    return false;
+  deleteClassroom(): boolean   {
+    if (confirm('Bent u zeker dat u deze klaslijst wilt verwijderen?')) {
+      this.classroomService.deleteClassroom(this.classroom).subscribe();
+    }
+    return true;
   }
